AbstractTrainer is created
